+++ src/features/reports/service.ts
@@ -1,6 +1,6 @@
 import { report } from "@/db/schema";
 import { reportRepository, reportService, uploadreport } from "@/types/reports"
-import { validatePostReport } from "@/utils/valueValidation";
+import { validateEditReport, validateId, validateNumberId, validatePostReport, validateReportLevel, validateReportType } from "@/utils/valueValidation";
 
 export function createReportService(repository:reportRepository):reportService {
 
@@ -10,22 +10,33 @@ export function createReportService(repository:reportRepository):reportService {
             return result
         },
         async getReportsWithType(type: string){
-             const result= await repository.getReportByType(type);
+            if (!validateReportType.safeParse(type)) return Promise.reject("Failed to validate report type.")
+
+            const result= await repository.getReportByType(type);
             return result
         },
          async getReportsWithlevel(level: number){
-             const result= await repository.getReportByLevel(level);
+            if (!validateReportLevel.safeParse(level)) return Promise.reject("Failed to validate report level.")
+
+            const result= await repository.getReportByLevel(level);
             return result
         },
          async getReportsWithTypeAndLevel(type: string,level:number){
-             const result= await repository.getReportByTypeAndLevel(type,level);
+            if (!validateReportType.safeParse(type)) return Promise.reject("Failed to validate report type.")
+            if (!validateReportLevel.safeParse(level)) return Promise.reject("Failed to validate report level.")
+
+            const result= await repository.getReportByTypeAndLevel(type,level);
             return result
         },
          async getReportWithId(id: number){
-             const result= await repository.getReportById(id);
+            if (!validateNumberId.safeParse(id)) return Promise.reject("Failed to validate report id.")
+
+            const result= await repository.getReportById(id);
             return result
         },
          async editReport(id: number,formdata:any){
+            if (!validateEditReport(id, formdata)) return Promise.reject("Failed to validate report.")
+
             const dataObject  = Object.fromEntries(formdata.entries());
             const data =dataObject as unknown as Partial<uploadreport>
             const result= await repository.editReport(id,data);
@@ -35,13 +46,14 @@ export function createReportService(repository:reportRepository):reportService {
             if (!validatePostReport(formdata)) return Promise.reject("Failed to validate report.")
 
             const dataObject  = Object.fromEntries(formdata.entries());
-            // implement zod here for it
             const data =dataObject as unknown as uploadreport
             data.createdAt =new Date().toUTCString()
             const result= await repository.createReport(data);
             return result
         },
         async deleteReport(id:number){
+            if (!validateNumberId.safeParse(id)) return Promise.reject("Failed to validate report id.")
+
             const result= await repository.deleteReportById(id)
             return result
         }
